delete alert

+++ src/Views/index.php
@@ -1,5 +1,4 @@
 <?php include __DIR__ . '/../views/partials/header.php'; ?>
-
 <div class="container my-5">
     <h2>Phone Book</h2>
     <br>
@@ -21,14 +20,13 @@
                     <td>
                         <img src="<?php echo '/phonebook/uploads/' . htmlspecialchars(basename($row['image'])); ?>" alt="profile image" style="width: 50px; height: 50px;">
                     </td>
-
                     <td><?php echo htmlspecialchars($row['name']); ?></td>
                     <td><?php echo htmlspecialchars($row['phone']); ?></td>
                     <td><?php echo htmlspecialchars($row['email']); ?></td>
                     <td><?php echo htmlspecialchars($row['category']); ?></td>
                     <td>
-                        <a class="btn btn-primary" href="index.php?action=edit&id=<?php echo $row['id']; ?>">Edit</a>
-                        <a class="btn btn-danger" href="index.php?action=delete&id=<?php echo $row['id']; ?>">Delete</a>
+                        <a class="btn btn-primary" href="index.php?action=edit&id=<?php echo $row['id']; ?>" onclick="confirmEdit();">Edit</a>
+                        <a class="btn btn-danger" href="index.php?action=delete&id=<?php echo $row['id']; ?>" onclick="confirmDelete();">Delete</a>
                     </td>
                 </tr>
             <?php endwhile; ?>
@@ -41,7 +39,6 @@
 
 </div>
 <script src="assets/js/bootstrap.min.js"></script>
-
 <script>
     document.addEventListener("DOMContentLoaded", () => {
         const rows = document.querySelectorAll(".clickable-row");
@@ -51,6 +48,15 @@
             });
         });
     });
+
+    function confirmEdit(){
+        return confirm("Are you sure you want to edit this contact.");
+        
+    }
+    function confirmDelete(){
+        return confirm("Are you sure you want to delete this contact.")? true: false;
+    }
+
 </script>
